requirements.txt
Django==4.2.7
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-countries==7.5.1
django-recaptcha==3.0.0
celery==5.3.4
redis==5.0.1
